Update README.md
The objective is to predict music genres based on a variety of features. This project was handled step by step with importing the dataset, data preprocessing, feature engineering using label encoder and standard scaler followed by dimensionality reduction by PCA. After this, the data was trained with Logistic Regression and evaluated using a classification matrix.